refactor(entity): Update package name to match master
