+++ src/main/java/com/kueerickpatrick/plantinfo/entity/plantObjects/PlantIndividualInfo.java
@@ -1,4 +1,4 @@
-package com.kueerickpatrick.plantinfo.entity.plantPOJO;
+package com.kueerickpatrick.plantinfo.entity.plantObjects;
 
 import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
